+++ libft.h
@@ -19,7 +19,7 @@
 # include <limits.h>
 
 int				ft_atoi(const char *str);
-void			ft_bzero(void *ptr, unsigned int n);
+void			ft_bzero(void *ptr, size_t n);
 void			*ft_calloc(size_t count, size_t size);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
@@ -29,9 +29,9 @@ int				ft_isprint(int c);
 char			*ft_itoa(int n);
 void			*ft_memchr(const void *str, int c, size_t n);
 int				ft_memcmp(const void *str1, const void *str2, size_t n);
-void			*ft_memcpy(void *dest, const void *src, unsigned int n);
-void			*ft_memmove(void *str1, const void *str2, unsigned int n);
-void			*ft_memset(void *ptr, int x, unsigned int n);
+void			*ft_memcpy(void *dest, const void *src, size_t n);
+void			*ft_memmove(void *str1, const void *str2, size_t n);
+void			*ft_memset(void *ptr, int x, size_t n);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putendl_fd(char *s, int fd);
 void			ft_putnbr_fd(int n, int fd);
@@ -41,8 +41,8 @@ char			*ft_strchr(const char *str, int c);
 char			*ft_strdup(const char *src);
 void			ft_striteri(char *s, void (*f)(unsigned int, char*));
 char			*ft_strjoin(char const *s1, char const *s2);
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
+unsigned int	ft_strlcat(char *dest, char *src, size_t size);
+unsigned int	ft_strlcpy(char *dest, char *src, size_t size);
 int				ft_strlen(const char *str);
 char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int				ft_strncmp(const char *str1, const char *str2, size_t n);
